Disabled reset on drop, build updates
For some reason, the code was working even with the "reset on drop" functionality enabled by default, which undoes every pull up attempt immediately afterwards.

This commit ensures that changes to the pin state are persisted properly and fixes the APIs that rely on changes being made to the GPIO pins.

Also amended some build details:
* Updated packages
* config.toml instead of config
* Self-hosting the toolchain for faster downloads and reliability
* Committed some RustRover detritus